fixup ci check
