+++ crates/rooch-framework-tests/src/tests/ethereum_validator_tests.rs
@@ -3,7 +3,8 @@
 
 use ethers::types::{Bytes, U256};
 use moveos_types::transaction::MoveAction;
-use rooch_key::keystore::{AccountKeystore, InMemKeystore};
+use rooch_key::keystore::account_keystore::AccountKeystore;
+use rooch_key::keystore::memory_keystore::InMemKeystore;
 use rooch_types::address::MultiChainAddress;
 use rooch_types::framework::empty::Empty;
